refactor: centralize color helpers
