+++ app/components/CategoryList.vue
@@ -128,6 +128,7 @@ import {
   deleteField
 } from 'firebase/firestore'
 import { getStorage, ref as storageRef, uploadBytes, deleteObject } from 'firebase/storage'
+import { textColor } from '../utils/color'
 
 interface Category {
   id?: string
@@ -330,23 +331,5 @@ const deleteCategory = async () => {
 const toggleFilter = (id?: string) => {
   if (!id) return
   activeCategoryId.value = activeCategoryId.value === id ? '' : id
-}
-
-function textColor(bg: string) {
-  const { r, g, b } = toRGB(bg);
-  const L = (0.2126 * r + 0.7152 * g + 0.0722 * b) / 255;
-  return L > 0.6 ? '#111827' : '#ffffff';
-}
-
-function toRGB(color: string) {
-  if (!color) return { r: 0, g: 0, b: 0 };
-  if (color.startsWith('#')) {
-    const hex = color.slice(1).replace(/^(.)(.)(.)$/, '$1$1$2$2$3$3');
-    const num = parseInt(hex, 16);
-    return { r: (num >> 16) & 255, g: (num >> 8) & 255, b: num & 255 };
-  }
-
-  const m = color.match(/\d+/g);
-  return m ? { r: +m[0], g: +Number(m[1]), b: +Number(m[2]) } : { r: 0, g: 0, b: 0 };
 }
 </script>
