+++ app/pages/share/[uid].vue
@@ -22,6 +22,7 @@ import { useFirebaseApp } from 'vuefire'
 import { getFirestore, doc, getDoc, collection, query, where, orderBy, getDocs } from 'firebase/firestore'
 import { ref, computed, onMounted } from 'vue'
 import { showError } from '#app'
+import { textColor } from '../../utils/color'
 
 interface Share { uid: string; date: string; categoryId: string }
 interface Todo { id?: string; title: string; done: boolean; categoryId: string | null; order: number; createdAt?: any }
@@ -73,21 +74,4 @@ onMounted(async () => {
     activeCategoryId.value = ''
   }
 })
-
-function textColor(bg: string) {
-  const { r, g, b } = toRGB(bg);
-  const L = (0.2126 * r + 0.7152 * g + 0.0722 * b) / 255;
-  return L > 0.6 ? '#111827' : '#ffffff';
-}
-
-function toRGB(color: string) {
-  if (!color) return { r: 0, g: 0, b: 0 };
-  if (color.startsWith('#')) {
-    const hex = color.slice(1).replace(/^(.)(.)(.)$/, '$1$1$2$2$3$3');
-    const num = parseInt(hex, 16);
-    return { r: (num >> 16) & 255, g: (num >> 8) & 255, b: num & 255 };
-  }
-  const m = color.match(/\d+/g);
-  return m ? { r: +m[0], g: +Number(m[1]), b: +Number(m[2]) } : { r: 0, g: 0, b: 0 };
-}
 </script>
